fix: fix errors for build
