+++ frontend/src/Router.tsx
@@ -7,7 +7,6 @@ import RegisterPage from './pages/RegisterPage'
 import { AccountPage } from './pages/AccountPage'
 import Index from './pages/Index'
 import UploadPage from './pages/UploadPage'
-import NewUploadPage from './pages/NewUploadPage'
 import { StudyDetailPage } from './pages/StudyDetailPage'
 import ProtectedRoute from './components/auth/ProtectedRoute'
 import PublicRoute from './components/auth/PublicRoute'
@@ -27,7 +26,7 @@ export function Router() {
         <Route path="/dashboard/account" element={<AccountPage />} />
         <Route path="/dashboard/study" element={<Index />} />
         <Route path="/dashboard/study/:id" element={<StudyDetailPage />} />
-        <Route path="/dashboard/uploadpage" element={<NewUploadPage />} />
+        <Route path="/dashboard/upload" element={<UploadPage />} />
       </Route>
       <Route
         path="/login"
@@ -43,14 +42,6 @@ export function Router() {
           <PublicRoute>
             <RegisterPage />
           </PublicRoute>
-        }
-      />
-      <Route
-        path="/upload"
-        element={
-          <ProtectedRoute>
-            <UploadPage />
-          </ProtectedRoute>
         }
       />
     </Routes>
