+++ frontend/src/components/ui/kibo-ui/ai/conversation.tsx
@@ -1,13 +1,13 @@
-'use client';
+'use client'
 
-import { Button } from '@repo/shadcn-ui/components/ui/button';
-import { ArrowDownIcon } from 'lucide-react';
-import type { ComponentProps } from 'react';
-import { useCallback } from 'react';
-import { StickToBottom, useStickToBottomContext } from 'use-stick-to-bottom';
-import { cn } from '@/lib/utils';
+import { Button } from '@/components/ui/button'
+import { ArrowDownIcon } from 'lucide-react'
+import type { ComponentProps } from 'react'
+import { useCallback } from 'react'
+import { StickToBottom, useStickToBottomContext } from 'use-stick-to-bottom'
+import { cn } from '@/lib/utils'
 
-export type AIConversationProps = ComponentProps<typeof StickToBottom>;
+export type AIConversationProps = ComponentProps<typeof StickToBottom>
 
 export const AIConversation = ({
   className,
@@ -20,25 +20,25 @@ export const AIConversation = ({
     role="log"
     {...props}
   />
-);
+)
 
 export type AIConversationContentProps = ComponentProps<
   typeof StickToBottom.Content
->;
+>
 
 export const AIConversationContent = ({
   className,
   ...props
 }: AIConversationContentProps) => (
   <StickToBottom.Content className={cn('p-4', className)} {...props} />
-);
+)
 
 export const AIConversationScrollButton = () => {
-  const { isAtBottom, scrollToBottom } = useStickToBottomContext();
+  const { isAtBottom, scrollToBottom } = useStickToBottomContext()
 
   const handleScrollToBottom = useCallback(() => {
-    scrollToBottom();
-  }, [scrollToBottom]);
+    scrollToBottom()
+  }, [scrollToBottom])
 
   return (
     !isAtBottom && (
@@ -52,5 +52,5 @@ export const AIConversationScrollButton = () => {
         <ArrowDownIcon className="size-4" />
       </Button>
     )
-  );
-};
+  )
+}
