+++ frontend/src/components/ui/kibo-ui/ai/message.tsx
@@ -1,14 +1,10 @@
-import {
-  Avatar,
-  AvatarFallback,
-  AvatarImage,
-} from '@repo/shadcn-ui/components/ui/avatar';
-import type { ComponentProps, HTMLAttributes } from 'react';
-import { cn } from '@/lib/utils';
+import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar'
+import type { ComponentProps, HTMLAttributes } from 'react'
+import { cn } from '@/lib/utils'
 
 export type AIMessageProps = HTMLAttributes<HTMLDivElement> & {
-  from: 'user' | 'assistant';
-};
+  from: 'user' | 'assistant'
+}
 
 export const AIMessage = ({ className, from, ...props }: AIMessageProps) => (
   <div
@@ -16,13 +12,13 @@ export const AIMessage = ({ className, from, ...props }: AIMessageProps) => (
       'group flex w-full items-end justify-end gap-2 py-4',
       from === 'user' ? 'is-user' : 'is-assistant flex-row-reverse justify-end',
       '[&>div]:max-w-[80%]',
-      className
+      className,
     )}
     {...props}
   />
-);
+)
 
-export type AIMessageContentProps = HTMLAttributes<HTMLDivElement>;
+export type AIMessageContentProps = HTMLAttributes<HTMLDivElement>
 
 export const AIMessageContent = ({
   children,
@@ -34,18 +30,18 @@ export const AIMessageContent = ({
       'flex flex-col gap-2 rounded-lg px-4 py-3 text-sm',
       'bg-muted text-foreground',
       'group-[.is-user]:bg-primary group-[.is-user]:text-primary-foreground',
-      className
+      className,
     )}
     {...props}
   >
     <div className="is-user:dark">{children}</div>
   </div>
-);
+)
 
 export type AIMessageAvatarProps = ComponentProps<typeof Avatar> & {
-  src: string;
-  name?: string;
-};
+  src: string
+  name?: string
+}
 
 export const AIMessageAvatar = ({
   src,
@@ -57,4 +53,4 @@ export const AIMessageAvatar = ({
     <AvatarImage alt="" className="mt-0 mb-0" src={src} />
     <AvatarFallback>{name?.slice(0, 2) || 'ME'}</AvatarFallback>
   </Avatar>
-);
+)
