+++ frontend/src/components/ui/kibo-ui/ai/reasoning.tsx
@@ -1,41 +1,41 @@
-'use client';
+'use client'
 
-import { useControllableState } from '@radix-ui/react-use-controllable-state';
+import { useControllableState } from '@radix-ui/react-use-controllable-state'
 import {
   Collapsible,
   CollapsibleContent,
   CollapsibleTrigger,
-} from '@repo/shadcn-ui/components/ui/collapsible';
-import { ChevronDownIcon } from 'lucide-react';
-import type { ComponentProps } from 'react';
-import { createContext, memo, useContext, useEffect, useState } from 'react';
-import { cn } from '@/lib/utils';
-import { AIResponse } from './response';
+} from '@/components/ui/collapsible'
+import { ChevronDownIcon } from 'lucide-react'
+import type { ComponentProps } from 'react'
+import { createContext, memo, useContext, useEffect, useState } from 'react'
+import { cn } from '@/lib/utils'
+import { AIResponse } from './response'
 
 type AIReasoningContextValue = {
-  isStreaming: boolean;
-  isOpen: boolean;
-  setIsOpen: (open: boolean) => void;
-  duration: number;
-};
+  isStreaming: boolean
+  isOpen: boolean
+  setIsOpen: (open: boolean) => void
+  duration: number
+}
 
-const AIReasoningContext = createContext<AIReasoningContextValue | null>(null);
+const AIReasoningContext = createContext<AIReasoningContextValue | null>(null)
 
 const useAIReasoning = () => {
-  const context = useContext(AIReasoningContext);
+  const context = useContext(AIReasoningContext)
   if (!context) {
-    throw new Error('AIReasoning components must be used within AIReasoning');
+    throw new Error('AIReasoning components must be used within AIReasoning')
   }
-  return context;
-};
+  return context
+}
 
 export type AIReasoningProps = ComponentProps<typeof Collapsible> & {
-  isStreaming?: boolean;
-  open?: boolean;
-  defaultOpen?: boolean;
-  onOpenChange?: (open: boolean) => void;
-  duration?: number;
-};
+  isStreaming?: boolean
+  open?: boolean
+  defaultOpen?: boolean
+  onOpenChange?: (open: boolean) => void
+  duration?: number
+}
 
 export const AIReasoning = memo(
   ({
@@ -52,44 +52,44 @@ export const AIReasoning = memo(
       prop: open,
       defaultProp: defaultOpen,
       onChange: onOpenChange,
-    });
+    })
     const [duration, setDuration] = useControllableState({
       prop: durationProp,
       defaultProp: 0,
-    });
+    })
 
-    const [hasAutoClosedRef, setHasAutoClosedRef] = useState(false);
-    const [startTime, setStartTime] = useState<number | null>(null);
+    const [hasAutoClosedRef, setHasAutoClosedRef] = useState(false)
+    const [startTime, setStartTime] = useState<number | null>(null)
 
     // Track duration when streaming starts and ends
     useEffect(() => {
       if (isStreaming) {
         if (startTime === null) {
-          setStartTime(Date.now());
+          setStartTime(Date.now())
         }
       } else if (startTime !== null) {
-        setDuration(Math.round((Date.now() - startTime) / 1000));
-        setStartTime(null);
+        setDuration(Math.round((Date.now() - startTime) / 1000))
+        setStartTime(null)
       }
-    }, [isStreaming, startTime, setDuration]);
+    }, [isStreaming, startTime, setDuration])
 
     // Auto-open when streaming starts, auto-close when streaming ends (once only)
     useEffect(() => {
       if (isStreaming && !isOpen) {
-        setIsOpen(true);
+        setIsOpen(true)
       } else if (!isStreaming && isOpen && !defaultOpen && !hasAutoClosedRef) {
         // Add a small delay before closing to allow user to see the content
         const timer = setTimeout(() => {
-          setIsOpen(false);
-          setHasAutoClosedRef(true);
-        }, 1000);
-        return () => clearTimeout(timer);
+          setIsOpen(false)
+          setHasAutoClosedRef(true)
+        }, 1000)
+        return () => clearTimeout(timer)
       }
-    }, [isStreaming, isOpen, defaultOpen, setIsOpen, hasAutoClosedRef]);
+    }, [isStreaming, isOpen, defaultOpen, setIsOpen, hasAutoClosedRef])
 
     const handleOpenChange = (open: boolean) => {
-      setIsOpen(open);
-    };
+      setIsOpen(open)
+    }
 
     return (
       <AIReasoningContext.Provider
@@ -104,15 +104,15 @@ export const AIReasoning = memo(
           {children}
         </Collapsible>
       </AIReasoningContext.Provider>
-    );
-  }
-);
+    )
+  },
+)
 
 export type AIReasoningTriggerProps = ComponentProps<
   typeof CollapsibleTrigger
 > & {
-  title?: string;
-};
+  title?: string
+}
 
 export const AIReasoningTrigger = memo(
   ({
@@ -121,13 +121,13 @@ export const AIReasoningTrigger = memo(
     children,
     ...props
   }: AIReasoningTriggerProps) => {
-    const { isStreaming, isOpen, duration } = useAIReasoning();
+    const { isStreaming, isOpen, duration } = useAIReasoning()
 
     return (
       <CollapsibleTrigger
         className={cn(
           'flex items-center gap-2 text-muted-foreground text-sm',
-          className
+          className,
         )}
         {...props}
       >
@@ -141,21 +141,21 @@ export const AIReasoningTrigger = memo(
             <ChevronDownIcon
               className={cn(
                 'size-4 text-muted-foreground transition-transform',
-                isOpen ? 'rotate-180' : 'rotate-0'
+                isOpen ? 'rotate-180' : 'rotate-0',
               )}
             />
           </>
         )}
       </CollapsibleTrigger>
-    );
-  }
-);
+    )
+  },
+)
 
 export type AIReasoningContentProps = ComponentProps<
   typeof CollapsibleContent
 > & {
-  children: string;
-};
+  children: string
+}
 
 export const AIReasoningContent = memo(
   ({ className, children, ...props }: AIReasoningContentProps) => (
@@ -165,9 +165,9 @@ export const AIReasoningContent = memo(
     >
       <AIResponse className="grid gap-2">{children}</AIResponse>
     </CollapsibleContent>
-  )
-);
+  ),
+)
 
-AIReasoning.displayName = 'AIReasoning';
-AIReasoningTrigger.displayName = 'AIReasoningTrigger';
-AIReasoningContent.displayName = 'AIReasoningContent';
+AIReasoning.displayName = 'AIReasoning'
+AIReasoningTrigger.displayName = 'AIReasoningTrigger'
+AIReasoningContent.displayName = 'AIReasoningContent'
